Move ControllerContext to a separate file
